refactor(windows): isolate proxy transaction models
